lexer: give `Span` a `FileId`, add a `SourceMap` to resolve it
`Span` carried start/end/line/col but no file identity, so a diagnostic
about two positions in two different files could only print two bare
`line:col` pairs — indistinguishable when they coincide (#75).

`Span` now carries a `FileId`: an opaque index into a `SourceMap`, which
owns the paths. `Lexer::new_in_file`/`tokenize_collecting_errors_in_file`
stamp one in; `Lexer::new` and friends keep producing `FileId::ANONYMOUS`,
which resolves to no path. `Span::locate` is the single place a position
is turned into displayable text — `path:line:col` when the file is
known, `line:col` when it isn't — so no diagnostic has to format
positions itself.

Byte offsets narrow from `usize` to `u32` (what rustc's own `BytePos`
does) to pay for the new field: `Span` sits inside every error enum, and
those enums are returned by value all through the pipeline under
clippy's `result_large_err` ceiling. At 20 bytes `Span` is smaller than
the 24 it was before. The lexer saturates rather than wrapping past
4 GiB of source.

Refs #75

+++ crates/hl-lexer/tests/lexer_tests.rs
@@ -389,7 +389,7 @@ fn eof_span_is_zero_width_at_end() {
     let eof = tokens.last().unwrap();
     assert_eq!(eof.kind, TokenKind::Eof);
     assert_eq!(eof.span.start, eof.span.end);
-    assert_eq!(eof.span.start, "a".len());
+    assert_eq!(eof.span.start as usize, "a".len());
 }
 
 #[test]
@@ -404,7 +404,10 @@ fn empty_input_eof_span_is_zero() {
 fn string_span_includes_quotes_but_lexeme_does_not() {
     let tok = single_token(r#""hello""#);
     assert_eq!(tok.lexeme.len(), 5); // "hello" without quotes
-    assert_eq!(tok.span.end - tok.span.start, tok.lexeme.len() + 2);
+    assert_eq!(
+        (tok.span.end - tok.span.start) as usize,
+        tok.lexeme.len() + 2
+    );
 }
 
 // --- iterator fusing behavior ---
